+++ clevertap-core/src/main/java/com/clevertap/android/sdk/BaseCallbackManager.java
@@ -2,7 +2,7 @@ package com.clevertap.android.sdk;
 
 import com.clevertap.android.sdk.displayunits.DisplayUnitListener;
 import com.clevertap.android.sdk.displayunits.model.CleverTapDisplayUnit;
-import com.clevertap.android.sdk.interfaces.DCDomainCallback;
+import com.clevertap.android.sdk.interfaces.SCDomainListener;
 import com.clevertap.android.sdk.interfaces.NotificationRenderedListener;
 import com.clevertap.android.sdk.interfaces.OnInitCleverTapIDListener;
 import com.clevertap.android.sdk.product_config.CTProductConfigListener;
@@ -22,7 +22,7 @@ public abstract class BaseCallbackManager {
 
     public abstract GeofenceCallback getGeofenceCallback();
 
-    public abstract DCDomainCallback getDCDomainCallback();
+    public abstract SCDomainListener getSCDomainListener();
 
     public abstract InAppNotificationButtonListener getInAppNotificationButtonListener();
 
@@ -53,7 +53,7 @@ public abstract class BaseCallbackManager {
 
     public abstract void setGeofenceCallback(GeofenceCallback geofenceCallback);
 
-    public abstract void setDCDomainCallback(DCDomainCallback dcDomainCallback);
+    public abstract void setSCDomainListener(SCDomainListener scDomainListener);
 
     public abstract void setInAppNotificationButtonListener(
             InAppNotificationButtonListener inAppNotificationButtonListener);